Update for Cypress 10
