+++ cypress/e2e/burger-store/local/creer-commande-mobile.feature
