+++ cypress/e2e/burger-store/local/creer-commande.spec.js
