+++ cypress/e2e/burger-store/pages/common.js
