+++ cypress/e2e/burger-store/pages/home.js
