+++ cypress/e2e/cypress/2-advanced-examples/actions.spec.js
