+++ cypress/e2e/cypress/2-advanced-examples/connectors.spec.js
