+++ cypress/e2e/cypress/2-advanced-examples/cookies.spec.js
