+++ cypress/e2e/cypress/2-advanced-examples/local_storage.spec.js
