+++ cypress/e2e/cypress/2-advanced-examples/location.spec.js
