+++ cypress/e2e/cypress/2-advanced-examples/window.spec.js
