+++ package.json
@@ -4,32 +4,33 @@
     "private": true,
     "scripts": {
         "kill": "pkill node",
-        "start":"npm run serve & npm run backend & npx wait-on http://localhost:8081",
+        "start": "npm run serve & npm run backend & npx wait-on http://localhost:8081",
         "serve": "vue-cli-service serve",
         "build": "vue-cli-service build",
         "backend": "json-server --watch db/db.json",
         "cypress:open": "npx cypress open",
         "cypress:run": "npx cypress run",
-        "cy2:run":"export CYPRESS_API_URL=http://localhost:1234 && npx cy2 run --record --key burger-store --parallel --ci-build `date +%s`"
+        "cy2:run": "export CYPRESS_API_URL=http://localhost:1234 && npx cy2 run --record --key burger-store --parallel --ci-build `date +%s`"
     },
     "dependencies": {
         "core-js": "^3.6.5",
-        "cypress": "^9.1.1",
+        "cypress": "^10.0.2",
         "cypress-cucumber-preprocessor": "^4.3.0",
         "cypress-multi-reporters": "^1.5.0",
-        "cypress-visual-regression": "file:cypress-visual-regression.tar",
         "json-server": "^0.17.0",
         "mocha-junit-reporter": "^2.0.2",
         "mochawesome": "^7.0.1",
-        "moment": "^2.29.1",
+        "moment": "^2.29.2",
         "shelljs": "^0.8.4",
         "vue": "^3.0.0",
-        "vue-router": "^4.0.0-0"
+        "vue-router": "^4.0.0-0",
+        "webpack": "4.46",
+        "webpack-dev-server": "^4.8.1"
     },
     "devDependencies": {
-        "@vue/cli-plugin-babel": "~4.5.0",
-        "@vue/cli-plugin-router": "~4.5.0",
-        "@vue/cli-service": "~4.5.0",
+        "@vue/cli-plugin-babel": "^4.5.17",
+        "@vue/cli-plugin-router": "^4.5.17",
+        "@vue/cli-service": "^4.5.17",
         "@vue/compiler-sfc": "^3.0.0",
         "cy2": "^1.3.0"
     },
